Add files via upload
Base POC BIO data request file with the addition of a geographical plotter of combined query results. The script contains a result counter to check whether results occur in all beacons. The results can then be plotted for a manual visual check. 'Species' is used to demonstrate functionality but this can be changed by the user.